feat: 타입 분리하여 구현
