+++ src/shared/types/FunnelType.ts
@@ -1,0 +1,21 @@
+import { FunnelProps, StepProps } from '../../features/event-manage/hooks/useFunnel';
+
+export enum StepNames {
+  HostSelection = 'HostSelection',
+  HostCreation = 'HostCreation',
+  EventTitle = 'EventTitle',
+  EventPeriod = 'EventPeriod',
+  EventOrganizerInfo = 'EventOrganizerInfo',
+  EventInfo = 'EventInfo',
+  EventType = 'EventType',
+  EventTag = 'EventTag',
+}
+
+export interface EventFunnelInterface {
+  onNext: (nextStep: string) => void;
+  onPrev: (prevStep: string) => void;
+  Funnel: React.FC<FunnelProps>;
+  Step: React.FC<StepProps>;
+  setStep: (step: number) => void;
+  currentStep: number;
+}
